pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "comparators"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "click",
    "pyyaml",
    "rocrate",
    "numpy",
    "pandas",
    "datacompy"
]

[project.scripts]
compare = "comparators.comparison_cli:compare"

[tool.setuptools.packages.find]
where = ["."]
include = ["comparators*"]